OV-52: * move OpenAi service
