+++ backend/src/bundles/chat/chat.controller.ts
@@ -9,12 +9,15 @@ import { ApiPath } from '~/common/enums/enums.js';
 import { HttpCode, HTTPMethod } from '~/common/http/http.js';
 import { type Logger } from '~/common/logger/logger.js';
 
+import { MAX_TOKEN } from '../../common/services/open-ai/libs/constants/max-token.constant.js';
+import {
+    ChatPath,
+    OpenAIRole,
+} from '../../common/services/open-ai/libs/enums/enums.js';
+import { type OpenAIService } from '../../common/services/open-ai/open-ai.service.js';
 import { type ChatService } from './char.service.js';
-import { MAX_TOKEN } from './libs/constants/max-token.constant.js';
-import { ChatPath, OpenAIRole } from './libs/enums/enums.js';
 import { type GenerateTextRequestDto } from './libs/types/types.js';
 import { textGenerationValidationSchema } from './libs/validation-schemas/validation-schemas.js';
-import { type OpenAIService } from './open-ai.service.js';
 
 class ChatController extends BaseController {
     private openAIService: OpenAIService;
