+++ backend/src/bundles/chat/libs/types/types.ts
@@ -1,4 +1,3 @@
 export { type ChatService } from './chat-service.type.js';
 export { type Message } from './message.type.js';
-export { type OpenAIService } from './open-ai-service.type.js';
 export { type GenerateTextRequestDto } from 'shared';
