+++ backend/src/common/services/open-ai/libs/enums/open-ai-role.enum.ts
