+++ backend/src/common/services/open-ai/libs/types/open-ai-service.type.ts
@@ -1,4 +1,4 @@
-import { type Message } from './message.type.js';
+import { type Message } from '~/bundles/chat/libs/types/types.js';
 
 type OpenAIService = {
     generateText(messages: Message[]): Promise<string>;
