+++ backend/src/common/services/open-ai/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type OpenAIService } from './open-ai-service.type.js';
