+++ backend/src/common/services/open-ai/open-ai.service.ts
@@ -2,13 +2,11 @@ import OpenAI, { type OpenAI as OpenAIType } from 'openai';
 
 import { type BaseConfig } from '~/common/config/base-config.package.js';
 
+import { type Message } from '../../../bundles/chat/libs/types/types.js';
 import { CHAT_MODEL, MAX_TOKEN } from './libs/constants/constants.js';
-import {
-    type Message,
-    type OpenAIService as OpenAIServiceModule,
-} from './libs/types/types.js';
+import { type OpenAIService as OpenAIServiceType } from './libs/types/types.js';
 
-class OpenAIService implements OpenAIServiceModule {
+class OpenAIService implements OpenAIServiceType {
     private config: BaseConfig;
     private openAi: OpenAIType;
